requirements.txt
autopep8==1.6.0
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
psycopg2-binary==2.9.3
python-decouple==3.6
model-mommy==2.0.0
